Gyrometer/viewer/requirements.txt
# Python dependencies for the orientation viewer
flask>=3.0.0
flask-socketio>=5.3.0
pyserial>=3.5
numpy>=1.24.0

# Optional: only needed for orientation_viewer.py (standalone 3D viewer)
# pygame>=2.5.0
# PyOpenGL>=3.1.7
# PyOpenGL-accelerate>=3.1.7
